RoundedRectangle/cornerRadius fixes
